Move to requestMainResource
